GPT_CALC_SERVER/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.37.13
botocore==1.37.13
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chess==1.11.2
click==8.0.3
cloud-init==25.1.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.3
django-storages==1.14.5
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mutagen==1.45.1
netifaces==0.11.0
numpy==2.2.5
oauthlib==3.2.0
openai==1.52.1
packaging==24.0
pgvector==0.4.0
pillow==10.2.0
psycopg==3.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodomex==3.11.0
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
python-apt==2.4.0+ubuntu4
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.1
pyxattr==0.7.2
PyYAML==5.4.1
requests==2.31.0
s3transfer==0.11.4
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
systemd-python==234
threadpoolctl==3.6.0
tqdm==4.66.5
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.2.1
wadllib==1.3.6
websockets==12.0
yt-dlp==2024.3.9.232714.dev0
zipp==1.0.0
zope.interface==5.4.0
